changed id type from Long to Integer

+++ src/main/java/com/michaeladrummonds/aguafina/services/CustomerService.java
@@ -14,5 +14,5 @@ public interface CustomerService {
 
     Customer updateCustomer(Customer customer);
 
-    Customer getCustomerById(Long id);
+    Customer getCustomerById(Integer id);
 }
